+++ src/main/java/joshie/harvest/core/helpers/SpawnItemHelper.java
@@ -59,10 +59,10 @@ public class SpawnItemHelper {
     public static void dropBlockAsItem(World world, int x, int y, int z, @Nonnull ItemStack stack) {
         if (!world.isRemote && world.getGameRules().getBoolean("doTileDrops") && !world.restoringBlockSnapshots) {
             float f = 0.7F;
-            double d0 = (double) (world.rand.nextFloat() * f) + (double) (1.0F - f) * 0.5D;
-            double d1 = (double) (world.rand.nextFloat() * f) + (double) (1.0F - f) * 0.5D;
-            double d2 = (double) (world.rand.nextFloat() * f) + (double) (1.0F - f) * 0.5D;
-            EntityItem entityitem = new EntityItem(world, (double) x + d0, (double) y + d1, (double) z + d2, stack);
+            double d0 = world.rand.nextFloat() * f + (1.0F - f) * 0.5D;
+            double d1 = world.rand.nextFloat() * f + (1.0F - f) * 0.5D;
+            double d2 = world.rand.nextFloat() * f + (1.0F - f) * 0.5D;
+            EntityItem entityitem = new EntityItem(world, x + d0, y + d1, z + d2, stack);
             entityitem.setPickupDelay(10);
             world.spawnEntity(entityitem);
         }
